Update Sensor_Data_Acquisition.ino
Code was updated to allow for sending information about the leak status to the surface.

The leak check was moved to the bottom of the main loop as the display code was created with it referenced as the last string.

The data.print statements were added to send the status to the surface. I do believe how I wrote the commas in allows for proper string separation?